feat : 과외일정 변경된 케이스 반영

+++ shared/scheduling-support/src/main/java/com/yedu/scheduling/support/ScheduleService.java
@@ -5,6 +5,7 @@ import java.time.ZoneId;
 import java.util.Date;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
 import org.quartz.Job;
 import org.quartz.JobBuilder;
 import org.quartz.JobDataMap;
@@ -16,6 +17,7 @@ import org.quartz.Trigger;
 import org.quartz.TriggerBuilder;
 import org.springframework.stereotype.Service;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class ScheduleService {
@@ -24,6 +26,8 @@ public class ScheduleService {
 
   @SneakyThrows
   public <T extends Job> void schedule(LocalDateTime executeTime, Class<T> jobClass, JobDataMap jobDataMap, String jobName)  {
+    JobKey jobKey = JobKey.jobKey(jobName);
+
     JobDetail jobDetail = JobBuilder.newJob(jobClass)
         .withIdentity(jobName)
         .usingJobData(jobDataMap)
@@ -31,21 +35,46 @@ public class ScheduleService {
         .storeDurably(true)
         .build();
 
+    if (scheduler.checkExists(jobKey)) {
+      log.warn("Job already exists: {}. Deleting existing job.", jobName);
+      scheduler.deleteJob(jobKey);
+    }
+
     Trigger trigger = TriggerBuilder.newTrigger()
         .withIdentity(jobName + "_trigger")
         .startAt(Date.from(executeTime.atZone(ZoneId.systemDefault()).toInstant()))
         .withSchedule(SimpleScheduleBuilder.simpleSchedule()
+            .withRepeatCount(0)
             .withMisfireHandlingInstructionFireNow()
         )
         .build();
 
     scheduler.scheduleJob(jobDetail, trigger);
+    log.info("Job scheduled: {} at {}", jobName, executeTime);
   }
 
   @SneakyThrows
   public <T extends Job> boolean cancelScheduledJob(String jobName) {
     JobKey jobKey = JobKey.jobKey(jobName);
-    return scheduler.deleteJob(jobKey);
+
+    if (scheduler.checkExists(jobKey)) {
+      log.warn("Job already exists: {}. Deleting existing job.", jobName);
+      scheduler.deleteJob(jobKey);
+    }
+
+    if (!scheduler.checkExists(jobKey)) {
+      log.warn("Job does not exist: {}", jobName);
+      return false;
+    }
+
+    boolean deleted = scheduler.deleteJob(jobKey);
+    if (deleted) {
+      log.info("Job cancelled: {}", jobName);
+      return true;
+    }
+
+    log.error("Failed to cancel job: {}", jobName);
+    return false;
   }
 
 }
